LayoutTests/fast/js/function-names-expected.txt

This test checks the names of functions constructed two different ways.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Function('1').toString().replace(/[ \n]+/g, ' ') is 'function anonymous() { 1; }'
PASS document.documentElement.onclick.toString().replace(/[ \n]+/g, ' ') is 'function onclick(event) { 2; }'
PASS successfullyParsed is true

TEST COMPLETE
